# api/asset/swagger.yaml
openapi: 3.0.0
info:
  title: "Starter Kit API Documentation"
  description: "sahakolay.com"
  version: "1.0.0"
components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      in: header
      name: Authorization
paths:
  /login:
    post:
      summary: "User Login"
      operationId: "UserLogin"
      tags:
        - "Auth"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
                  format: password
      responses:
        200:
          description: "Success login"
          content:
            application/json:
              example:
                status: true
                message: "Success"
                data:
                  {
                    "token": "eyJhbGciOiJIUzI.eyJpc3MiOiIxIiwiZXhwI.yyMRyoRa1H8sjqwajdVUyK",
                  }
  /register:
    post:
      summary: "User Create"
      operationId: "UserCreate"
      tags:
        - "Auth"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fullname
                - email
                - password
                - phone
              properties:
                fullname:
                  type: string
                email:
                  type: string
                password:
                  type: string
                  format: password
                phone:
                  type: string
      responses:
        200:
          description: "User Create"
          content:
            application/json:
              example:
                status: true
                message: "Success"
                data:
                  {
                    "token": "eyJhbGciOiJIUzI.eyJpc3MiOiIxIiwiZXhwI.yyMRyoRa1H8sjqwajdVUyK",
                    "user":
                      {
                        "id": 1,
                        "fullname": "John DOE",
                        "email": "johndone@gmail.com",
                        "phone": "5432222222",
                        "is_admin": false,
                      },
                  }